Fix Cluster wheel to include new enums
The Cluster objects use the new enum module. Make sure that the enum
module is part of the Clusters wheel instead of the Core wheel.

While at it, also fix Nullable type to be usable as an initializer,
which makes the Clusters library Python 3.11 compatible without touching
all the Cluster objects itself.

Also drop mobly dependency, which avoids a dependency to psutil. The
psutil dependency uses native code, which is especially annoying on
aarch64 platforms as there are no pre-built wheels available for aarch64
on PyPI. So this change will make the add-on Dockerfile simpler.