Make sure the generated version is Python safe
Currently dev builds have a zero prependet (e.g. dev0102). Python seems
to drop that 0 when passing to the build system, leading to a wheel with
a version without leading zero (dev101).

Don't use the leading zero to make it a valid Python version from the
beginning and avoid unexpected file names later on during the build.