Add and use local CHIP project config headers
Instead of relying on the in-tree config headers in config/standalone/
pass a path to the project config to let the framework use our own
config header files. This avoids using patches to adjust build specific
configs.

Note that previously the builds in this repository used config/standalone/
instead of config/python/. The CHIP project config headers in config/python/
is what the script ./scripts/build_python.sh uses. This script is sometimes
used to build the Python library during development. And was the reason
why the wrong file got patched in #79.

There are minor differences between the standalone and python specific
configs, specifically in CHIPProjectConfig.h. This PR adds a header file
based on the config in config/python/, but with small adjustments.

Compared to the previous configuration, the effective change at build
time should work out to:
- Dropping device specific configs (CHIP_DEVICE_CONFIG_*). These should
  not have an effect for our controller builds.
- Increae of CHIP_CONFIG_CONTROLLER_MAX_ACTIVE_DEVICES to 1024 (this
  time the new configruation gets applied correclty).
- Increase of CHIP_CONFIG_MAX_EXCHANGE_CONTEXTS from 150 (standalone
  config) to 256 which feels more natural.