Add raw attribute callback support
This modifies the SDK to add support for untyped attribute update
callbacks. This allows to process updates of custom clusters and avoid
unnecessary type instantiation server side.

The change in the upstream SDK probably needs a bit more cleanup first.

This adds the same patch to the main branch as added in #68 to the release
branch. With this main branch builds can be used by the Python Server
again.