Disable host unit test config
By default the Python controller gets built with unit test configuration
enabled. This adds extra functionality to inject errors etc. for
testing. There is no known negative effect to have it enabled, but it is
certainly better to disable it for a production setup.

Furthermore this fixes error handling for APIs with callbacks and
improves logging somewhat.

Specifically, this integrates changes from the following PRs
- https://github.com/project-chip/connectedhomeip/pull/34346
- https://github.com/project-chip/connectedhomeip/pull/34354
- https://github.com/project-chip/connectedhomeip/pull/34368